+++ parse.js
@@ -1,6 +1,7 @@
 import fs from 'fs';
 import { Parser } from '@syuilo/aiscript';
+import { inspect } from 'util';
 
 const script = fs.readFileSync('./test.is', 'utf8');
 const ast = Parser.parse(script);
-console.log(JSON.stringify(ast, null, 2));
+console.log(inspect(ast, { depth: 10 }));