+++ src/error.ts
@@ -1,7 +1,10 @@
+import type { Loc } from './node.js';
+
 export abstract class AiScriptError extends Error {
 	// name is read by Error.prototype.toString
 	public name = 'AiScript';
 	public info?: any;
+	public loc?: Loc;
 
 	constructor(message: string, info?: any) {
 		super(message);
@@ -30,8 +33,8 @@ export class NonAiScriptError extends AiScriptError {
  */
 export class AiScriptSyntaxError extends AiScriptError {
 	public name = 'Syntax';
-	constructor(message: string, info?: any) {
-		super(message, info);
+	constructor(message: string, public loc: Loc, info?: any) {
+		super(`${message} (Line ${loc.line}, Column ${loc.column})`, info);
 	}
 }
 /**
@@ -39,8 +42,18 @@ export class AiScriptSyntaxError extends AiScriptError {
  */ 
 export class AiScriptTypeError extends AiScriptError {
 	public name = 'Type';
-	constructor(message: string, info?: any) {
-		super(message, info);
+	constructor(message: string, public loc: Loc, info?: any) {
+		super(`${message} (Line ${loc.line}, Column ${loc.column})`, info);
+	}
+}
+
+/**
+ * Namespace collection errors.
+ */
+export class AiScriptNamespaceError extends AiScriptError {
+	public name = 'Namespace';
+	constructor(message: string, public loc: Loc, info?: any) {
+		super(`${message} (Line ${loc.line}, Column ${loc.column})`, info);
 	}
 }